requirements.txt
numpy 
scikit-image
matplotlib
scikit-learn
torch
torchvision
opencv-python
opencv-python-headless 
opencv-contrib-python
pillow
tensorflow